refactor: move message_bus to top-level messaging module
Move AsyncMessageBus out of fold_db_core::infrastructure into a new
top-level crate::messaging module so both schema (domain) and
fold_db_core (coordinator) can depend on it without a circular
reference. The previous location implied message_bus was infrastructure
OF fold_db_core, but it is actually a cross-cutting concern used by
both layers.

- Add pub mod messaging in lib.rs
- Rename src/fold_db_core/infrastructure/message_bus/ -> src/messaging/
- fold_db_core::infrastructure re-exports AsyncMessageBus/Event from
  crate::messaging for call-site compatibility
- Update all imports across schema, fold_db_core, tests, testing_utils

No behavior change; no API change visible outside the crate except the
new module path.
